chore: move reload logic from cmd/run to a shared extension runner helper
